requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
flask>=2.0.1
